+++ packages/agentmark/tsup.config.ts
@@ -2,11 +2,21 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: ['src/index.ts'],
+  format: ['cjs', 'esm'],
+  dts: true,
   splitting: false,
   sourcemap: true,
   clean: true,
-  dts: true,
-  format: ['esm', 'cjs'],
-  minify: false,
   target: 'es2019',
+  treeshake: true,
+  noExternal: ['@puzzlet/templatedx'],
+  outExtension({ format }) {
+    return {
+      js: format === 'cjs' ? '.cjs' : '.js'
+    }
+  },
+  esbuildOptions(options) {
+    options.mainFields = ['module', 'main']
+    options.platform = 'node'
+  },
 });